Wrapped serialize and deserialize in classes to allow extension
